dodaj adminove ovalsti i harkodiraj trenera

+++ Mindfulness.Server/Controllers/ChallengeController.cs
@@ -86,11 +86,25 @@ public class ChallengeController(MindfulnessDbContext context, IMapper mapper) :
     [HttpDelete("{id:guid}")]
     public async Task<IActionResult> DeleteChallenge(Guid id)
     {
-        var challenge = await context.Challenges.FindAsync(id);
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+        if (userId is null)
+        {
+            return BadRequest("User not found");
+        }
         
+        var userGuid = Guid.Parse(userId);
+        
+        var challenge = await context.Challenges.FirstOrDefaultAsync(x => x.Id == id);
+
         if (challenge is null)
         {
             return NotFound();
+        }
+
+        if (challenge.UserId != Guid.Parse(userId) && !User.IsInRole("Admin"))
+        {
+            return Unauthorized();
         }
         
         context.Challenges.Remove(challenge);
